# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Northstar treasury APIs",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8181"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
    "/api/accounts": {
      "get": {
        "summary": "List up to 25 own accounts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Missing operator role"
          },
          "404": {
            "description": "Account absent or owned by another identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Rejected monetary operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached"
          }
        }
      },
      "post": {
        "summary": "Open a zero-balance account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccount"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Missing operator role"
          },
          "404": {
            "description": "Account absent or owned by another identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Rejected monetary operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached"
          }
        }
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "summary": "Read an owned account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Missing operator role"
          },
          "404": {
            "description": "Account absent or owned by another identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Rejected monetary operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached"
          }
        }
      }
    },
    "/api/accounts/{id}/movements": {
      "get": {
        "summary": "Read last 50 movements in descending sequence",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movement"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Missing operator role"
          },
          "404": {
            "description": "Account absent or owned by another identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Rejected monetary operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached"
          }
        }
      },
      "post": {
        "summary": "Record an immutable movement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostMovement"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created, or original movement on identical retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movement"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Missing operator role"
          },
          "404": {
            "description": "Account absent or owned by another identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Rejected monetary operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached"
          }
        }
      }
    },
    "/api/audit": {
      "get": {
        "summary": "Read latest 100 own audit events; eventually consistent",
        "responses": {
          "200": {
            "description": "Events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "event_id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "account_id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "amount": {
                        "type": "string",
                        "pattern": "^-?[0-9]+(\\.[0-9]{1,2})?$",
                        "description": "Exact decimal string; never convert monetary values to binary floating point."
                      },
                      "balance": {
                        "type": "string",
                        "pattern": "^-?[0-9]+(\\.[0-9]{1,2})?$",
                        "description": "Exact decimal string; never convert monetary values to binary floating point."
                      },
                      "sequence": {
                        "type": "integer"
                      },
                      "occurred_at": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Missing operator role"
          },
          "404": {
            "description": "Account absent or owned by another identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Concurrent update or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Rejected monetary operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit reached"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Account": {
        "type": "object",
        "required": [
          "id",
          "name",
          "currency",
          "balance",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "enum": [
              "USD",
              "EUR"
            ]
          },
          "balance": {
            "type": "string",
            "pattern": "^-?[0-9]+(\\.[0-9]{1,2})?$",
            "description": "Exact decimal string; never convert monetary values to binary floating point."
          },
          "version": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Movement": {
        "type": "object",
        "required": [
          "id",
          "accountId",
          "amount",
          "balance",
          "description",
          "occurredAt",
          "sequence"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "string",
            "pattern": "^-?[0-9]+(\\.[0-9]{1,2})?$",
            "description": "Exact decimal string; never convert monetary values to binary floating point."
          },
          "balance": {
            "type": "string",
            "pattern": "^-?[0-9]+(\\.[0-9]{1,2})?$",
            "description": "Exact decimal string; never convert monetary values to binary floating point."
          },
          "description": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "sequence": {
            "type": "integer"
          }
        }
      },
      "CreateAccount": {
        "type": "object",
        "required": [
          "name",
          "currency"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "currency": {
            "type": "string",
            "enum": [
              "USD",
              "EUR"
            ]
          }
        }
      },
      "PostMovement": {
        "type": "object",
        "required": [
          "amount",
          "description"
        ],
        "properties": {
          "amount": {
            "oneOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+(\\.[0-9]{1,2})?$"
              }
            ]
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 140
          }
        }
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        }
      }
    }
  }
}